Move handlers to subdirectory
