+++ src/handlers/HandlerError.js
@@ -1,4 +1,4 @@
-const BaseError = require('./BaseError');
+const BaseError = require('../BaseError');
 
 class HandlerError extends BaseError {
     constructor(message) {
